Added logic to detect disappearance of parent.
This logic is a bit pro-tem. In theory tmon should
parse stdin using katcl, and fail if katcl returns
end of file. However, wrinkle is that stdin may not be
the same as stdout, when used interactively. However,
when used as kcs we are pretty sure that these are the same,
problem is that it is unclear if we should rely on this ...