Create /bin/kcpfpg symlink on kcpfpg install
tcpborphserver3 relies upon the kcpfpg utility being available as
`/bin/kcpfpg`, but the kcpfpg Makefile installs it into `$(PREFIX)/bin`.
Since the #define in the tcpborphserver3 code cannot predict what the
user might use for PREFIX, I decided to keep the #define unchanged (for
backwards compatibility) and just create `/bin/kcpfpg` as a symlink to
`$(PREFIX)/bin/kcpfpg`.  This will break if the user sets PREFIX to an
empty string or `/` since it will try to create a symlink to itself.